Fix notes
